add `types.rs` file
keep types used in multiple systems in the same file
